Move SparseInferenceRewriteAction to a dedicated file
